gobject-introspection: Introduce GIR search path.
Allow to set a NIX_GIR_PATH environment variable which can be used to set a
search PATH for .gir files, where consecutive paths are separated by colons.

The implementation in compiler.c is very ugly and far from DRY, but I'm not an
expert with glib. So don't hesitate to improve things here if you have the
patience I definitely didn't have.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>