unity: Add Unity core package and dependencies.
Well, what a mess! Finally it's building successfully and even managed to do
some cleanups.

Probably most people might be wondering why I'm doing a port of Unity to Nix
while I personally don't use and/or even like it. The main reason is because I
tried to introduce someone to Nix and NixOS who actively uses Unity and doesn't
want to switch to something more sane^H^H^H^Hlightweight.

So let me introduce him: @gradel

He will be the new maintainer of the Unity desktop packages in Nix and I'm going
to mentor him during this process.

Enough about introduction, there still is a lot of work to do:

 * Improve directory structure of pkgs/desktop/unity.
 * Add more packages to propagatedBuildInputs to keep things DRY.
 * Maybe clean up my GIR compiler patch.
 * Add an update script to update all that stuff at once.
 * Separate the ugly sed hacks into nice clean patches.
 * Check build warnings and maybe fix them instead of removing -Werror.
 * Add missing dependencies.
 * Add missing meta information (I pretty much omitted about all of them).
 * The packages ibus, libgee and PackageKit obviously don't belong to unity,
   move them to the right locations.

And the biggest TODO:

Actually *test* it, as it is only UnityCore, I presume there is quite a bunch of
packages missing at this point.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>