Add applyAbiCompatiblePatches function.
This function is in charge of compiling security updates.  It does so by
taking advantage for the purity of the Nix language.  It does not change any
of the hashes, even when an identical set of packages is given as argument
for the quickfixPackages.

The purity of the Nix language offer us a clean way to reason about Nixpkgs
as if this was a simple function.  Thus, the following property hold true,
even when applied to a collection of packages:

  let f = nixpkgsFun; in
    fix f == f (f (fix f));

This property is used for building security updates, by using a variant of
the first function:

  let f = nixpkgsFun; g = fixpkgsFun; in
     g (fix f)

Using a fixpkgsFun function which is a function similar to nixpkgsFun, we
can generate a similar set of packages while only recompiling the fixed
packages.

The problem then becomes patching any package which are dependening on the
fixed packages. This is what the applyAbiCompatiblePatches functions do.

      +-------+
      |       |
  +---v----+  |  +--------+      +--------+
  |        |  |  |        |      |        |
  |  pkgs  +-----> onefix +-+  +-+ recfix |
  |        |     |        | |  | |        |
  +--------+     +--------+ |  | +-----^--+
                            |  |       |
                         +--v--v--+    |
                         |        |    |
                         | abifix +----+
                         |        |
                         +--------+

The loop back on pkgs is the original fix-point on the default set of
packages.  Then we recompile the patched programs, in the one-fix function
call. The rec-fix is a way to probe for recompilations, by comparing the
dependencies of one-fix packages, taken from pkgs, with the dependencies of
rec-fix packages, taken from the abi-fix packages. The abi-fix packages is a
zipping function which takes one-fix packages and rec-fix packages for
patching one-fix packages.  In practice, it also takes the original set of
packages, to ensure that packages have the same package name length, and
that other which depend on them can be patched safely.